🎨 refactor(examples/with-redux): overhaul after moving to app router
