+++ examples/with-redux/jest.config.ts
@@ -9,9 +9,7 @@ const jestConfig: Config = {
   extensionsToTreatAsEsm: ['.ts', '.tsx'],
   testEnvironment: 'jsdom',
   setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-  moduleNameMapper: {
-    '@/(.*)': '<rootDir>/src/$1',
-  },
+  moduleNameMapper: { '@/(.*)': '<rootDir>/src/$1' },
   transform: {
     '^.+\\.(t|j)sx?$': '@swc/jest',
     '.+\\.(css|styl|less|sass|scss)$': 'jest-css-modules-transform',
