+++ examples/with-redux/src/app/components/Counter/counter.module.scss
