+++ examples/with-redux/src/app/components/Nav.tsx
@@ -1,11 +1,21 @@
+'use client'
+
 /* Core */
 import Link from 'next/link'
+import { usePathname } from 'next/navigation'
+import cx from 'classnames'
 
 export const Nav = () => {
+  const pathname = usePathname()
+
   return (
-    <section>
-      <Link href="/">Home</Link>
-      <Link href="/verify">Verify page</Link>
-    </section>
+    <nav>
+      <Link className={cx({ active: pathname === '/' })} href="/">
+        Home
+      </Link>
+      <Link className={cx({ active: pathname === '/verify' })} href="/verify">
+        Verify
+      </Link>
+    </nav>
   )
 }
