+++ examples/with-redux/src/app/layout.tsx
@@ -1,14 +1,65 @@
-/* Instruments */
-import './styles/globals.css'
-
 /* Components */
-import { Providers } from './providers'
+import { Nav } from './components'
+import { Providers } from '@/lib/providers'
+
+/* Instruments */
+import styles from './styles/main.module.scss'
+import './styles/globals.scss'
 
 export default function RootLayout(props: React.PropsWithChildren) {
   return (
     <Providers>
       <html lang="en">
-        <body>{props.children}</body>
+        <body>
+          <section className={styles.container}>
+            <Nav />
+
+            <header>
+              <img src="/logo.svg" className={styles.logo} alt="logo" />
+            </header>
+
+            <main>{props.children}</main>
+
+            <footer>
+              <span>Learn </span>
+              <a
+                className={styles.link}
+                href="https://reactjs.org/"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                React
+              </a>
+              <span>, </span>
+              <a
+                className={styles.link}
+                href="https://redux.js.org/"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                Redux
+              </a>
+              <span>, </span>
+              <a
+                className={styles.link}
+                href="https://redux-toolkit.js.org/"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                Redux Toolkit
+              </a>
+              ,<span> and </span>
+              <a
+                className={styles.link}
+                href="https://react-redux.js.org/"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                React Redux
+              </a>
+            </footer>
+          </section>
+        </body>
       </html>
     </Providers>
   )
