--- examples/with-redux/src/app/providers.tsx
@@ -1,16 +1,0 @@
-'use client'
-
-/* Core */
-import { Provider } from 'react-redux'
-
-/* Instruments */
-import { reduxStore } from '@/lib/redux'
-
-export const Providers = (props: React.PropsWithChildren) => {
-  return (
-    <Provider store={reduxStore}>
-      {/* @ts-expect-error reason: react-redux and next.js type mismatch https://github.com/vercel/next.js/issues/37421 */}
-      {props.children}
-    </Provider>
-  )
-}
