+++ examples/with-redux/src/app/styles/globals.scss
@@ -1,5 +1,6 @@
 html,
 body {
+  min-height: 100vh;
   padding: 0;
   margin: 0;
   font-family: system-ui, sans-serif;
